[Fix] Fix databricks_cluster_pluginframework data source (#4097)
## Changes
After
https://github.com/databricks/terraform-provider-databricks/commit/bac842d603465ddbf3582e2dab2d9189c17e3113
, the autogenerated structures used for interacting with
state/config/plan always use Lists for nested structures, even if there
is only ever at most one, for compatibility with older versions of the
TF provider. The cluster data source on the plugin framework is
handwritten and also needs to be converted to use lists instead of a
pointer as well.

## Tests
Ran the `TestAccDataSourceClusterByID` and
`TestAccDataSourceClusterByName` integration tests, which now pass.

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK