[Fix] Mark unity_catalog_provisioning_state as ReadOnly (#4116)
## Changes
<!-- Summary of your changes that are easy to understand -->
In a recent update to the GoSDK, a new field called
`unity_catalog_provisioning_state` was added. This field changes values
based on server decisions, so it's supposed to be marked as computed.
This pull request sorts out that issue.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>