[Feature] Add `databricks_query` resource instead of `databricks_sql_query` (#4103)
## Changes
<!-- Summary of your changes that are easy to understand -->

This PR is built on top of #4051 which should be merged first.

The new resource uses the new [Queries
API](https://docs.databricks.com/api/workspace/queries/create) instead
of the legacy one that will be deprecated. Since the new resource has a
slightly different set of parameters, it was decided to create a new
resource and deprecate the old one.

This resource uses old TF SDK to be compatible with TF exporter (until
#4050 is implemented).

TODOs:

- Need to discuss how to handle permissions - `sql_query` permissions
look like working, but not sure if we should continue to use that API
- Support in the exporter will be in a separate PR


## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [x] using Go SDK