[Exporter] Improving reliability of `Emit` function
I found in the large-scale testing that sometimes we don't handle emitting of the same
resource reliably, and this may lead to generation of duplicate resources (very small
amount, but still) - found this in a very specific case when notebooks were listed without
directories.

This PR fixes this problem:

- by tracking if resource is already in importing queue
- detecting duplicates during code generation

It also may improve performance a bit (2-3%).