[Exporter] Improving reliability of `Emit` function (#4163)
## Changes
<!-- Summary of your changes that are easy to understand -->

I found in the large-scale testing that sometimes we don't handle
emitting of the same resource reliably, and this may lead to generation
of duplicate resources (very small amount, but still) - found this in a
very specific case when notebooks were listed without directories.

This PR fixes this problem:

- by tracking if resource is already in importing queue
- detecting duplicates during code generation

It also may improve performance a bit (2-3%).

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK