[Internal] Migrate Share Resource to Plugin Framework (#4047)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR migrates the share resource to the Plugin framework. The code
was largely copied "as is" from the previous implementation of the share
resource, with the necessary adaptations made for integration with the
Plugin framework.

This implementation utilizes the newly generated Effective fields to
provide the functionality that was previously achieved through diff
suppression.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [x] using Go SDK

---------

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>