[Fix] Handle edge case for `effective_properties` in `databricks_sql_table` (#4153)
## Changes
<!-- Summary of your changes that are easy to understand -->

It was reported in #4098 that some of the specified options, like,
`multiLine`, `recursiveFileLookup` and potentially more, aren't returned
as `option.multiLine`, etc., but instead are expanded into full names,
like, `spark.sql.dataSourceOptions.multiLine`.

This PR changes lookup logic a bit, and if we can't find
`option.something`, then we're looking for all options ending with
`.something` (only if there are no `.` in the name).

Resolves #4098

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK