[Internal] Migrate Share Data Source to Plugin Framework (#4161)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR migrates the share/shares data sources to the Plugin framework.
The code was largely copied "as is" from the previous implementation of
the share data source, with the necessary adaptations made for
integration with the Plugin framework.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
~~Note: current tests create shares using the SDKv2 resource, but fetch
them using the new plugin framework data source. Once the resource
migration will be merged, I will amend this.~~ Edit: Now that the
resource itself is merged, the acceptance tests use the plugin
framework's version of the resource.

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [x] using Go SDK

---------

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>