[Internal] Use isServiceProposedIfEmpty annotations for effective fields (#4270)
This PR is a follow-up to #4057 and #4112. We're shifting away from
using the "computed" annotation to decide if a field should have
"effective field" behavior. Instead, we're adopting the more precise
"isServiceProposedIfEmpty" annotation. This lines up better with what we
aim to achieve with effective fields, without making all regular
computed and non-effective fields generate effective field code.

We'll only merge this once the relevant fields are updated accordingly
in the OpenAPI spec.

## Changes
<!-- Summary of your changes that are easy to understand -->

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [ ] using Go SDK

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>