[Internal] Generate effective fields based of isServiceProposedIfEmpty (#4282)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR introduces a change that uses the new isServiceProposedIfEmpty
annotation to generate the TFSDK structs. With this update, only the
fields with this annotation will have an EffectiveField combo.
Previously, fields that were marked as Computed would automatically get
this combo, but that's no longer the case. As a result, most of the
EffectiveField combos are now removed.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>