[Exporter] better handling of online tables/vsis in listing (#4288)
## Changes
<!-- Summary of your changes that are easy to understand -->

Right now, to export vector search indexes and online tables, the
`uc-tables` needs to be specified in the listing. This PR removes this
limitation by checking if the corresponding service is enabled without
requiring `uc-tables` to be enabled as well.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK