[Internal] Move TFSDK model template to universe (#4303)
## Changes
<!-- Summary of your changes that are easy to understand -->
In order to get ready for auto-generating resources, we're relocating
the templates for our plugin framework models to an internal repository.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
Verified that generating an SDK before and after this change resulted in
a no-op.

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>