[Internal] Panic if the provided path is invalid (#4309)
## Changes
Callers of `CustomizableSchema`'s methods can supply invalid paths.
Currently, this returns an error, but this is never OK and should always
be fixed before making any changes to a resource. Here, I change
`navigateSchemaWithCallback` to panic if the supplied path does not
exist.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK