[Internal] Use Attributes by default for List Objects (#4315)
## Changes
Resources implemented on the SDKv2 used blocks to represent two kinds of
fields: traditional lists, and nested complex types (which were treated
as a list of length at most 1). The Plugin Framework now has support for
list attributes, and it is recommended to use this in general, except
for places where compatibility with SDKv2 is needed.

To achieve this, this change ensures that *StructToSchema methods in the
plugin framework generate attributes for lists by default. Then, for
resources migrated from SDKv2, the implementer must call
`ConfigureForSdkV2Migration()` in the customize callback. This will
convert any list attributes into list blocks, preserving compatibility
with SDKv2-implemented resources.

Additionally, this PR improves the handling of computed, non-optional
fields. For lists annotated as `tf:"computed"`, the resulting schema is
invalid: it is marked as computed and as required because of the absence
of the `"optional"` tag. When a field is labeled as computed, it can be
marked as optional but should never be marked as required; that is
addressed here as well.

Finally, this PR also includes some autogenerated changes modifying Apps
to use `types.Object` in place of `types.List` for nested complex types.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK