[Internal] Generate both SdkV2-compatible and Plugin Framework-compatible structures (#4332)
## Changes
To make it possible to implement new resources following the Plugin
Framework's recommendation to use attributes while allowing resources to
be migrated from SDKv2 to the Plugin Framework without breaking, we will
create two copies of each code-generated structure in the Terraform
provider. The structures with no suffix should be used for new
resources. The structures with the `_SdkV2` suffix should be used for
resources in the plugin framework that are migrated from SDKv2.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework