[Fix] Send only what is required in Update of `databricks_credential` (#4349)
## Changes
<!-- Summary of your changes that are easy to understand -->

This is a workaround for a problem in the API spec - right now, the
Update operation tries to send the full struct, although the API accepts
only one attribute of that struct. For GCP we should omit the struct
completely...

The problem doesn't exist for `storage_credential` on AWS as it uses the
correct structure for the Update request, but problem exists for GCP.

Resolves #4335

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] using Go SDK
- [ ] using TF Plugin Framework