[Feature] Skip calling Read after Create/Update operations for notebooks (#4173)
## Changes
<!-- Summary of your changes that are easy to understand -->

It's based on #4190.

It was found that the `import` API returns `object_id` as a result of
its execution, so we don't really need to call the `Read` operation that
will call `get-status` to fill all attributes (when we create a notebook
using the `SOURCE` format).

The other formats (`DBC` and `JUPYTER`) will still call the `Read` to
fill in the missing parameters, such as `language`.

It also changed the `language` to `optional,computed` so we can safely
set the value and don't trigger the diff.

This should help when we import a large number of notebooks, i.e., when
applying exported resources.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] ~relevant change in `docs/` folder~
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [ ] ~using Go SDK~

---------

Co-authored-by: Omer Lachish <289488+rauchy@users.noreply.github.com>