[Feature] Add `no_compute` attribute to `databricks_app` (#4364)
## Changes
Support `no_compute` flag for Apps. The flag is only used on create to
determine whether compute should be provisioned as part of the Create
App RPC. Changes to this flag for existing resources are no-ops and
simply propagated from plan to state.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: Tanmay Rustagi <tanmay.rustagi@databricks.com>
Co-authored-by: Tanmay Rustagi <88379306+tanmay-db@users.noreply.github.com>