[Doc] Officially deprecate `databricks_table` and add migration instructions (#4463)
## Changes
Some customers are confused about the inconsistent state of the
`databricks_table` resource. It is not documented, but it is also not
deprecated officially. This PR clarifies this by finally marking the
resource as deprecated, as it has been replaced by
`databricks_sql_table`. The documentation explains how to migrate from
the old resource to the new.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework