[Fix] Do not send `run_as` when updating `databricks_pipeline` when not specified (#4470)
## Changes
A recent change to add support for `run_as` in pipelines causes pipeline
update requests to include the `run_as` field. The Get Pipeline always
returns the `run_as_user_name` for a pipeline, even for workspaces not
in the private preview. Some users may have `run_as` in their TF state
now, even if they did not specify it in their configuration. As we are
using SDKv2 for this resource, it isn't straightforward to see whether
the user actually included `run_as` in their configuration or if it is
only in the state because of the change to the `Read` method.

This PR addresses this by only sending `run_as` if there is a planned
change in the `run_as` field. If the user doesn't set `run_as`
explicitly (as in existing customers not in the private preview), or if
it is set to the same value as the `run_as_user_name` field, the
provider does not need to set `run_as`, so it can be removed from the
request.

Resolves #4469.

## Tests

- [x] Created & updated a pipeline in a workspace without run_as preview
enabled; succeeded.
- [x] Created & updated a pipeline in a workspace with run_as preview
enabled; succeeded.