[Fix] Increase `databricks_library` installation timeout from 15m to 30m (#4479)
## Changes
Currently, `databricks_library` waits for 15m for libraries to install.
We received a report that this was a bit too short for some use cases.
For now, I'll bump this up to 30m to mitigate the issue faced by this
customer.

The long-term plan to fix this should be based on adopting
https://developer.hashicorp.com/terraform/plugin/framework/resources/timeouts
in our provider.

## Tests

I did not do any tests for this change.