[Fix] Share Reordering Issue (#4481)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR fixes an issue where reordering objects in a (pluginfw) Share
would fail unless other changes were made to the objects. The bug caused
the new order to be ignored if no additional modifications were present.
With this fix, reordering now correctly updates the Share, whether or
not other changes are made.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

Co-authored-by: Omer Lachish <rauchy@users.noreply.github.com>