[Exporter] Explicitly abort execution via `panic` if list of users can't be fetched (#4500)
## Changes
<!-- Summary of your changes that are easy to understand -->

The list of users is used for different purposes, like checking if
workspace objects (notebooks/directories/files/...) should be exported
or not. So it makes sense to abort execution if SCIM backend returns us
an error, otherwise we'll mark all workspace objects as ignored.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework