[Exporter] Fix matching on `workspace_path` and refactoring
During optimization of workspace objects handling we started to generate WS resources
directly, but forgot to set properties like `workspace_path` that are used in the
matching, so no references were generated (not critical, but still incorrect).  It did
happen only when notebooks/directories were listed in the same execution.

Also refactored code a bit to avoid code duplicates, and improve matching.