[Fix] Only allow `authorized_paths` to be updated in the `options` field of `databricks_catalog` (#4517)
## Changes
There was a misunderstanding with the Unity Catalog API in connection
with
https://github.com/databricks/terraform-provider-databricks/pull/4476.
Apparently, only `authorized_paths` is eligible for update in the UC
Catalog API. Otherwise, any other changes to `options` must trigger
resource recreation.

This PR customizes the logic of `databricks_catalog` to do this. To be
specific:
* `CustomizeDiff` marks the `options` field as `ForceNew` if there is a
key other than `authorized_paths` which was updated.
* The `Update` command removes any options that aren't
`authorized_paths`. This might be the case in the future if other
options are added for HMS catalogs that can be set at create time but
not updated.
* The `Create` command removes any options from the `UpdateCatalog` RPC.

## Tests
Added two integration tests to verify the behavior of
`databricks_catalog` update.
* Updated `TestUcAccCatalogUpdate` to verify that adding options with a
key other than `authorized_paths` to a catalog defined without any
options causes the catalog to be recreated.
* Added `TestUcAccCatalogHmsConnectionUpdate` which tests the lifecycle
of HMS catalogs, including updating the `authorized_paths` option,
enforcing that the catalog resource is never deleted during the
lifecycle of the test.

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>
Co-authored-by: vuong-nguyen <44292934+nkvuong@users.noreply.github.com>