[Fix] Remove configuration drift when configuring `databricks_connection` to built-in Hive Metastore (#4505)
## Changes
<!-- Summary of your changes that are easy to understand -->

When configuring built-in Hive Metastore, the backend returns a number
of additional options. But it's hard to suppress them as they are valid
when `builtin = "false"`, and each option is evaluated separately. So
it's easier just to remove them before saving to the state.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: vuong-nguyen <44292934+nkvuong@users.noreply.github.com>