[Doc] Added permissions for GCP file events (#4415)
## Changes
Added permissions for GCP file events. As it stands, this is a docs only
change, similar to [the PR for
Azure](https://github.com/databricks/terraform-provider-databricks/pull/4412).
I'm wondering if it's possible and/or desirable to wrap the created role
into a custom terraform resource as is the case with AWS. Whatever we
end up with, I'll need to e2e test this before merging.

## Tests

- [ ] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: vuong-nguyen <44292934+nkvuong@users.noreply.github.com>
Co-authored-by: Alex Ott <alexey.ott@databricks.com>