[Exporter[ Fix panic caused by incorrect values in the cluster policies
The cluster policy is supposed to use `string` type for `value` and `defaultValue` fields
of the cluster policy, so explicit cast was done in the code. But it caused the crash when
reading an incorrectly specific cluster policy (the user error).  This PR checks if
interface type is correct before casting