databricks_mount: replace listing of mounted directories with a check if the directory exists
This directory listings can take very long when there are lots of files in the mounted directory. This slows down both the state refresh and the actual mounting up to the point that the operation times out.
Checking if the director exists fails if the directory is not readable eg because of invalid credentials