[Exporter] Fix panic caused by incorrect values in the cluster policies (#4585)
## Changes
<!-- Summary of your changes that are easy to understand -->

The cluster policy is supposed to use `string` type for `value` and
`defaultValue` fields of the cluster policy, so explicit cast was done
in the code. But it caused the crash when reading an incorrectly
specific cluster policy (the user error). This PR checks if interface
type is correct before casting

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework