[Fix] `databricks_mount`: remove redundant listing of mounted directories (#4590)
Directory listings can take very long when there are lots of files in
the mounted directory. This slows down both the state refresh and the
actual mounting up to the point that the operation times out.
Fixes
[4509](https://github.com/databricks/terraform-provider-databricks/issues/4509)

## Changes
<!-- Summary of your changes that are easy to understand -->
Remove directory listing of mounted directories.


## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>