Fix databricks_workspace_binding for existing resources using `external-location` or `storage-credential` (#4611)
## Changes
Previously, the UC workspace binding API allowed users to specify
`external-location` and `storage-credential` for the securable type in
the [update securable workspace bindings
API](https://docs.databricks.com/api/workspace/workspacebindings/updatebindings).
This was changed to `external_location` and `storage_credential`.
However, for users who created these bindings using the older names,
those names were persisted in the `id` of the resource. As a result, the
securable type is always updated to the older name, causing a permanent
drift to appear.

This PR changes the `Read()` method of `databricks_workspace_binding` to
rewrite old-style securable types in this API to their new style names.
The result is that there is no longer a difference in the
`securable_type` attribute, which is the expected behavior.

## Tests
Added a unit test to verify that an ID in the old style would be
interpreted appropriately by the updated read method, resulting in no
diff when the ID contains `external-location` and the `securable_type`
is set to `external_location`.

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>